Refactor CloudResrouceMover in its own file
